backend/service/requirements.txt
anyio==3.6.2 ; 
astroid==2.12.12 ; 
asyncpg==0.27.0 ; 
attrs==22.1.0 ; 
certifi==2022.9.24 ; 
click==8.1.3 ; 
colorama==0.4.6 ; 
dill==0.3.6 ; 
elastic-transport==8.4.0 ; 
elasticsearch==8.5.0 ; 
exceptiongroup==1.0.4 ; 
fastapi==0.87.0 ; 
greenlet==2.0.1 ; 
h11==0.14.0 ; 
httpcore==0.16.1 ; 
httptools==0.5.0 ; 
httpx==0.23.1 ; 
idna==3.4 ; 
iniconfig==1.1.1 ; 
lazy-object-proxy==1.8.0 ; 
mako==1.2.4 ; 
markupsafe==2.1.1 ; 
mccabe==0.7.0 ; 
mypy-extensions==0.4.3 ; 
packaging==21.3 ; 
pathspec==0.10.2 ; 
platformdirs==2.5.4 ; 
pluggy==1.0.0 ; 
psycopg2-binary==2.9.5 ; 
pydantic==1.10.2 ; 
pyparsing==3.0.9 ; 
pytest-asyncio==0.20.2 ; 
pytest==7.2.0 ; 
python-dotenv==0.21.0 ; 
python-version==0.0.2 ; 
pyyaml==6.0 ; 
rfc3986[idna2008]==1.5.0 ; 
sniffio==1.3.0 ; 
starlette==0.21.0 ; 
tomli==2.0.1 ; 
tomlkit==0.11.6 ; 
typing-extensions==4.4.0 ; 
urllib3==1.26.12 ; 
uvicorn[standard]==0.20.0 ; 
uvloop==0.17.0 ; 
watchfiles==0.18.1 ; 
websockets==10.4 ; 
wrapt==1.14.1 ; 
